+++ shepherd_data/shepherd_data/ivonne.py
@@ -19,10 +19,9 @@ from types import TracebackType
 
 import numpy as np
 import pandas as pd
+from shepherd_core.writer import Writer as CoreWriter
 from tqdm import trange
 from typing_extensions import Self
-
-from shepherd_core.writer import Writer as CoreWriter
 
 from .mppt import MPPTracker
 from .mppt import OptimalTracker