+++ shepherd_data/shepherd_data/reader.py
@@ -9,14 +9,14 @@ from pathlib import Path
 import h5py
 import numpy as np
 from matplotlib import pyplot as plt
+from shepherd_core.data_models import EnergyDType
+from shepherd_core.logger import get_verbose_level
+from shepherd_core.logger import log
 from tqdm import trange
 
 from shepherd_core import Reader as CoreReader
 from shepherd_core import Writer as CoreWriter
 from shepherd_core import local_tz
-from shepherd_core.data_models import EnergyDType
-from shepherd_core.logger import get_verbose_level
-from shepherd_core.logger import log
 
 # import samplerate  # noqa: ERA001, TODO: just a test-fn for now